Added checking of chunks to improve DX in error cases.
When there is some issue in the Babel configuration e.g. missing comments in source files after transpile, the current behavior fails without reporting any error. In my opinion it seems safe to add a test for all passed over `chunkNames` and validate whether they exist in the assets as well. If not, report an error.